Modify package.json and unit-tests
Added tasks to the package for running mocha and coverage locally
