+++ .gitignore
@@ -2,3 +2,4 @@ node_modules/
 coverage/
 *.swp
 *.swo
+npm-debug.log
