+++ README.md
@@ -7,19 +7,17 @@ A repository for testing MochaJS, chai and chai-as-promises.
 
 1- git clone git@github.com:facundovictor/mocha-testing.git
 
-2- npm install -g mocha
-
 3- npm install -g istanbul
 
 3- npm install (for installing package devDependencies)
 
 ## To run the tests
 
-Just type **mocha** or **npm run**
+Just type **npm test**
 
 ## To get the coverage
 
-```istanbul cover _mocha```
+Just type **npm run coverage**
 
 This should generate the coverage folder.
 To see the results, just open **./coverage/lcov-report/index.html**
