+++ test/unit/users.test.js
@@ -1,3 +1,14 @@
+var sinon = require('sinon');
+var chai = require("chai");
+var chaiAsPromised = require("chai-as-promised");
+chai.use(chaiAsPromised);
+var expect = require('chai').expect;
+var jsf = require('json-schema-faker');
+var mocks = require('../mocks');
+var controllers = require('../../controllers');
+var should = chai.should();
+var promise = require('bluebird');
+
 describe('Users', function() {
 
   var user, model, ctrl;
@@ -7,11 +18,11 @@ describe('Users', function() {
 
     user = jsf(mocks.user);
 
-    // Let's suppose that we have these tow functions in the "model".
-    user['getAPIInformation'] = function(callback) {
+    // Let's suppose that we have these two functions in the "model".
+    user.getAPIInformation = function(callback) {
       callback(1,2,3);
     };
-    user['saveAPIInformation'] = function(info) {
+    user.saveAPIInformation = function(info) {
       /* Let's suppose that we save some data */
       return {
         response: 200
@@ -64,7 +75,7 @@ describe('Users', function() {
   context('When editing an user', function() {
 
     beforeEach(function(){
-      user['save'] = sandbox.stub();
+      user.save = sandbox.stub();
       user.save.withArgs().returns(user);
     });
 
@@ -108,7 +119,7 @@ describe('Users', function() {
 
   context('When getting an user', function() {
     beforeEach(function(){
-      user['getBestFriend'] = sandbox.stub();
+      user.getBestFriend = sandbox.stub();
       user.getBestFriend.withArgs().returns(promise.resolve(jsf(mocks.user)));
     });
 
@@ -168,7 +179,7 @@ describe('Users', function() {
 
     context(", An error occurs", function() {
       beforeEach(function(){
-        user['saveAPIInformation'] = sandbox.stub();
+        user.saveAPIInformation = sandbox.stub();
         user.saveAPIInformation.withArgs().returns({
           response: 404
         });
